gemini/requirements.txt
fastapi==0.115.9
google-generativeai==0.8.4
Pillow==11.1.0
python-dotenv==1.0.1
python-multipart==0.0.20
google-api-core==2.24.1
requests==2.31.0
xmltodict==0.14.2
uvicorn == 0.34.0
colorlog == 6.9.0